+++ pages/user/login.tsx
@@ -13,8 +13,8 @@ import GoogleBtn from '@components/auth/googleBtn';
 import Button from '@components/common/Button';
 import ErrorMessage from '@components/common/ErrorMessage';
 import Input from '@components/common/Input';
+import InputPassword from '@components/common/InputPassword';
 import HeadMeta from '@components/headMeta';
-import InputPassword from '@components/inputPassword';
 import NATURE_IMAGE from '@public/user-login.jpg';
 
 interface ILoginProps {
@@ -231,7 +231,7 @@ export default function Login() {
                     message: '최소 8개 이상의 문자 포함해야 합니다.',
                   },
                 })}
-              ></InputPassword>
+              />
               {isWrong === true && data?.ok === false && (
                 <ErrorMessage text={data?.message} />
               )}
